do_the_shuffle
Fix nullable build errors.
